uclient-fetch: Support of --method, --body-data and --body-file
The --method allows executing PUT, DELETE, OPTIONS, HEAD e.g.:

    wget -O - -q \
    --method=PUT \
    --body-data='{"id": 42}' \
    --header='Content-Type: application/json' \
    http://api.example.com/

To avoid clashes with the --post-data/file options,
the separate --body-data/file must be used.
They are stored to the same post_data and post_file vars.
Mutual usage of the options is not allowed, same as in GNU wget.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>
Signed-off-by: Felix Fietkau <nbd@nbd.name>